python-ai/requirements.txt
flask
flask-cors
requests
python-dotenv
requests-html
beautifulsoup4
lxml
